Revert "Move search bar to core"
